+++ epyk/__init__.py
@@ -1,9 +1,11 @@
-"""
-
-"""
-
 from epyk.core import Page as Rpt
 from epyk.core.data import events
+from epyk.core.html import Defaults as settings
+
 from epyk.core.css.styles.attributes import CssInline
+
+# Add JavaScript shortcuts
+from epyk.core.js import std
+from epyk.core.js import expr
 
 Page = Rpt.Report